requirements.txt
flask>=3.0
gunicorn>=21.0
PyPDF2>=3.0
pymupdf>=1.23
pdf2docx>=0.5
python-docx>=1.1
openpyxl>=3.1
reportlab>=4.0
werkzeug>=3.0
pdfplumber>=0.10
python-dotenv>=1.0
# Para validación MIME type
# En Windows: pip install python-magic-bin
# En Linux/Docker: pip install python-magic
python-magic>=0.4.24
